Added groundwork for Jump and Speed evolutions
Remember to put correct tag to the ground object for the script to correctly update the touchingGround boolean.
Added booleans to control which evolution the player chose.
Need to check if jumping script actually works. 
Evolutions will have to be changed by the script controlling the Hub, most likely.
Consider if need to DontDestroyOnLoad in Awake ().